[FLINK-7057][tests] extract FailingBlockingInvokable from CoordinatorShutdownTest
